class decrypter & resource renamer

+++ native_class_encrypter/src/class_encrypter.h
@@ -6,7 +6,7 @@ extern "C" {
 JNIEXPORT jbyteArray JNICALL Java_dev_sharkuscator_obfuscator_encryption_ClassEncrypter_encrypt(JNIEnv*, jclass, jbyteArray, jbyteArray);
 }
 
-void xxtea_encrypt_class(size_t, unsigned char*, const unsigned char*);
+void xxtea_encrypt_class(jint, unsigned char*, const unsigned char*);
 
 void xxtea_encrypt_block(int, unsigned char*, const unsigned char*);
